+++ pkg/scheduler/framework/session_plugins.go
@@ -4,6 +4,7 @@
 package framework
 
 import (
+	"maps"
 	"net/http"
 
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/api"
@@ -76,6 +77,10 @@ func (ssn *Session) AddPreemptScenarioValidatorFn(rf api.ScenarioValidatorFn) {
 
 func (ssn *Session) AddReclaimVictimFilterFn(rf api.VictimFilterFn) {
 	ssn.ReclaimVictimFilterFns = append(ssn.ReclaimVictimFilterFns, rf)
+}
+
+func (ssn *Session) AddBindRequestMutateFn(fn api.BindRequestMutateFn) {
+	ssn.BindRequestMutateFns = append(ssn.BindRequestMutateFns, fn)
 }
 
 func (ssn *Session) CanReclaimResources(reclaimer *podgroup_info.PodGroupInfo) bool {
@@ -340,4 +345,12 @@ func (ssn *Session) NodeOrderFn(task *pod_info.PodInfo, node *node_info.NodeInfo
 
 func (ssn *Session) IsRestrictNodeSchedulingEnabled() bool {
 	return ssn.SchedulerParams.RestrictSchedulingNodes
+}
+
+func (ssn *Session) MutateBindRequestAnnotations(pod *pod_info.PodInfo, nodeName string) map[string]string {
+	annotations := map[string]string{}
+	for _, fn := range ssn.BindRequestMutateFns {
+		maps.Copy(fn(pod, nodeName), annotations)
+	}
+	return annotations
 }