+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - Added optional pod and namespace label selectors to limit the scope of monitored pods
-
+- Added a plugin extension point for scheduler plugins to add annotations to BindRequests.
 
 ## [v0.6.0] - 2025-06-16